% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRecMotif-methods.R
\name{accessors}
\alias{accessors}
\alias{get_probe_set}
\alias{get_pbm_condition}
\alias{get_zscore_motif}
\alias{get_delta_zscore_motif}
\alias{get_array_id}
\alias{get_motif_strength}
\alias{get_rolling_ic}
\alias{get_seed_zscore}
\alias{get_seed_sequence}
\alias{get_beta}
\alias{get_motif}
\alias{get_motif_name}
\alias{get_icm}
\alias{get_pwm}
\alias{get_ppm}
\alias{get_match_motif}
\alias{get_match_name}
\alias{get_match_altname}
\alias{get_match_icm}
\alias{get_match_pwm}
\alias{get_match_ppm}
\alias{get_match_pvalue}
\alias{get_match_cluster}
\alias{set_probe_set}
\alias{set_pbm_condition}
\alias{set_zscore_motif}
\alias{set_array_id}
\alias{set_seed_sequence}
\alias{set_motif_name}
\title{Access or replace data from CoRecMotifs}
\usage{
get_probe_set(corecmotif)

get_pbm_condition(corecmotif)

get_zscore_motif(corecmotif)

get_delta_zscore_motif(corecmotif)

get_array_id(corecmotif)

get_motif_strength(corecmotif)

get_rolling_ic(corecmotif)

get_seed_zscore(corecmotif)

get_seed_sequence(corecmotif)

get_beta(corecmotif)

get_motif(corecmotif)

get_motif_name(corecmotif)

get_icm(corecmotif)

get_pwm(corecmotif)

get_ppm(corecmotif)

get_match_motif(corecmotif)

get_match_name(corecmotif)

get_match_altname(corecmotif)

get_match_icm(corecmotif)

get_match_pwm(corecmotif)

get_match_ppm(corecmotif)

get_match_pvalue(corecmotif)

get_match_cluster(corecmotif)

set_probe_set(corecmotif, value)

set_pbm_condition(corecmotif, value)

set_zscore_motif(corecmotif, value)

set_array_id(corecmotif, value)

set_seed_sequence(corecmotif, value)

set_motif_name(corecmotif, value)
}
\arguments{
\item{corecmotif}{\link[=CoRecMotif-class]{CoRecMotif}. The motif whose slots
should be accessed or replaced.}

\item{value}{Object to replace the slot with. The required type depends on
which slot is being replaced. See \linkS4class{CoRecMotif} for more information
on the expected classes.}
}
\value{
The value of the selected slot OR a \link[=CoRecMotif-class]{CoRecMotif}
with the appropriate slot updated.
}
\description{
Extracts or replaces data stored in \link[=CoRecMotif-class]{CoRecMotifs}.
}
