% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_matrix.R
\name{annotate_fluorescence_matrix}
\alias{annotate_fluorescence_matrix}
\title{Annotate fluorescence matrix}
\usage{
annotate_fluorescence_matrix(fluorescence_matrix, annotation)
}
\arguments{
\item{fluorescence_matrix}{A data frame of fluorescence values, such as
output by read_fluorescence_matrices().}

\item{annotation}{A data frame of PBM probe annotations. See 'Details' for
expected columns.}
}
\value{
A data frame of fluorescence values and annotations for each probe.
}
\description{
Adds probe information to a data frame of fluorescence values. See 'Details'
for a description of the inputs, including required annotation columns.
}
\details{
The input fluorescence matrix is expected to be of the form output by
read_fluorescence_matrices(), with rows corresponding to probes and columns
corresponding to PBM conditions. See \code{\link{read_fluorescence_matrices}}
for a more detailed description of the expected format.

The input annotation table must contain the columns \code{probeID},
\code{probe_or}, \code{probe_repl}, \code{probe_type}, ... The expected
content of these columns is described in detail below.

\describe{

\item{\code{probeID}}{The probe ID. This column will be used to map the probe
annotations to the values in the fluorescence matrix, so the IDs in this
column must match the IDs in the probeID column in the fluorescence matrix.
IDs that are not present in both will be dropped.}

\item{\code{probe_or}}{The orientation of the probe relative to the slide.
This column should contain the values "o1" and "o2". Only "o1" rows will be
used.}

\item{\code{probe_repl}}{The replicate number of the probe. This column
should contain the values "r1", "r2", etc. Only "r1" rows will be used.}

\item{\code{probe_type}}{The probe type. This column should contain the
values "MOTIF" and "BACKGROUND".}

}
}
\examples{
annotated_fluorescence_matrix <- annotate_fluorescence_matrix(
    fluorescence_matrix,
    annotation
)
}
