% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_annotate_fluorescence_table.R
\name{annotate_fluorescence_table}
\alias{annotate_fluorescence_table}
\title{Import and annotate fluorescence data}
\usage{
annotate_fluorescence_table(
  fluorescence_file,
  pbm_conditions,
  annotation,
  output_file = NULL
)
}
\arguments{
\item{fluorescence_file}{the path to the file containing the fluorescence
data to load. See 'Details' for expected columns.}

\item{pbm_conditions}{a character vector specifying the PBM conditions (e.g.,
cell type, treatment, and factor profiled) in the order they appear in
\code{fluorescence_file}.}

\item{annotation}{a data frame containing the probe annotations to use. See
\link{hTF_v1_annotation} for expected columns.}

\item{output_file}{the path to the TSV file where the annotated fluorescence
table will be written. If NULL, no file is created. (Default: NULL)}
}
\value{
A data frame containing the fluorescence values from the specified
fluorescence data file and the probe annotations from the specified
annotation file. See 'Details' for a description of each column.
}
\description{
Reads a file containing fluorescence data (in which rows correspond to probes
and columns correspond to fluorescence values) and adds probe metadata from
an annotation table.
}
\details{
\subsection{Fluorescence File}{

The fluorescence data file supplied to \code{fluorescence_file} must contain
\eqn{n + 3} tab-delimited columns, where n is the number of PBM conditions
supplied to \code{pbm_conditions}. This file must NOT have a header. The columns
must be as follows:
\enumerate{
\item The probe ID. The orientation tags (i.e., "_o1" or "_o2") will be removed
from these IDs.
\item The probe sequence.
\item The number of PBM conditions profiled on this array.
}

Columns 4 through \eqn{(n + 3)} must contain the fluorescence values for the
PBM conditions in the order they appear in \code{pbm_conditions}.
}

\subsection{Annotation}{

The IDs in the \code{probe_id} column must match the IDs in the first column of
the fluorescence data file (after the orientation tag has been removed). IDs
that are not present in both will be dropped.The output data frame contains
\eqn{6 + n} columns, where n is again the number of PBM conditions supplied
to the \code{pbm_conditions} parameter. The first six columns will be the columns
from the annotation table (see \link{hTF_v1_annotation} for expected columns). The
remaining columns will contain the fluorescence values for each condition and
will be named based on the vector of PBM conditions supplied to
\code{pbm_conditions}.
}
}
\examples{
print("FILL THIS IN")
}
