% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_check_replicates.R
\name{check_replicates}
\alias{check_replicates}
\title{Check reproducibility of a list of CoRecMotifs}
\usage{
check_replicates(
  corecmotifs,
  n_replicates = 2,
  eucl_distance = 0.4,
  output_file = NULL
)
}
\arguments{
\item{corecmotifs}{the list of \link[=CoRecMotif-class]{CoRecMotifs} to filter.}

\item{n_replicates}{a single positive integer specifying the minimum number
of replicates to require.}

\item{eucl_distance}{a single number specifying the maximum allowable
Euclidean distance between replicate motifs or NULL to skip the replicate
comparison step.}

\item{output_file}{the path to the RDS file where the list of reproducible
\link[=CoRecMotif-class]{CoRecMotifs} will be written. If NULL (the default), no
file is written.}
}
\value{
A list of \link[=CoRecMotif-class]{CoRecMotifs} that are reproducible.
}
\description{
Filter a list of \link[=CoRecMotif-class]{CoRecMotifs} based on whether they are
reproducible, i.e., found in multiple replicate experiments.
}
\examples{
print("FILL THIS IN")
}
