% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_conditions.R
\name{compare_conditions}
\alias{compare_conditions}
\title{Compare CoRecMotifs across different PBM conditions}
\usage{
compare_conditions(corecmotifs)
}
\arguments{
\item{corecmotifs}{\code{list}. The \link[=CoRecMotif-class]{CoRecMotifs} to compare to
each other.}
}
\value{
A data frame with comparison information about a list of
\link[=CoRecMotif-class]{CoRecMotifs}.
}
\description{
Compares the \link[=CoRecMotif-class]{CoRecMotifs} in each PBM condition to the
CoRecMotifs from the same probe set in all the other PBM conditions.
}
\examples{
print("FILL THIS IN")
}
