% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRecMotif-class.R, R/CoRecMotif-methods.R
\docType{class}
\name{CoRecMotif-class}
\alias{CoRecMotif-class}
\alias{probe_set,CoRecMotif-method}
\alias{pbm_condition,CoRecMotif-method}
\alias{zscore_motif,CoRecMotif-method}
\alias{delta_zscore_motif,CoRecMotif-method}
\alias{array_id,CoRecMotif-method}
\alias{motif_strength,CoRecMotif-method}
\alias{rolling_ic,CoRecMotif-method}
\alias{seed_zscore,CoRecMotif-method}
\alias{seed_sequence,CoRecMotif-method}
\alias{beta,CoRecMotif-method}
\alias{motif,CoRecMotif-method}
\alias{motif_name,CoRecMotif-method}
\alias{icm,CoRecMotif-method}
\alias{pwm,CoRecMotif-method}
\alias{ppm,CoRecMotif-method}
\alias{match_motif,CoRecMotif-method}
\alias{match_name,CoRecMotif-method}
\alias{match_altname,CoRecMotif-method}
\alias{match_icm,CoRecMotif-method}
\alias{match_pwm,CoRecMotif-method}
\alias{match_ppm,CoRecMotif-method}
\alias{match_pvalue,CoRecMotif-method}
\alias{match_cluster,CoRecMotif-method}
\alias{probe_set<-,CoRecMotif-method}
\alias{pbm_condition<-,CoRecMotif-method}
\alias{zscore_motif<-,CoRecMotif-method}
\alias{array_id<-,CoRecMotif-method}
\alias{seed_sequence<-,CoRecMotif-method}
\alias{motif_name<-,CoRecMotif-method}
\alias{as.data.frame,CoRecMotif-method}
\alias{show,CoRecMotif-method}
\title{CoRecMotif class}
\usage{
\S4method{probe_set}{CoRecMotif}(x)

\S4method{pbm_condition}{CoRecMotif}(x)

\S4method{zscore_motif}{CoRecMotif}(x)

\S4method{delta_zscore_motif}{CoRecMotif}(x)

\S4method{array_id}{CoRecMotif}(x)

\S4method{motif_strength}{CoRecMotif}(x)

\S4method{rolling_ic}{CoRecMotif}(x)

\S4method{seed_zscore}{CoRecMotif}(x)

\S4method{seed_sequence}{CoRecMotif}(x)

\S4method{beta}{CoRecMotif}(x)

\S4method{motif}{CoRecMotif}(x)

\S4method{motif_name}{CoRecMotif}(x)

\S4method{icm}{CoRecMotif}(x)

\S4method{pwm}{CoRecMotif}(x)

\S4method{ppm}{CoRecMotif}(x)

\S4method{match_motif}{CoRecMotif}(x)

\S4method{match_name}{CoRecMotif}(x)

\S4method{match_altname}{CoRecMotif}(x)

\S4method{match_icm}{CoRecMotif}(x)

\S4method{match_pwm}{CoRecMotif}(x)

\S4method{match_ppm}{CoRecMotif}(x)

\S4method{match_pvalue}{CoRecMotif}(x)

\S4method{match_cluster}{CoRecMotif}(x)

\S4method{probe_set}{CoRecMotif}(x) <- value

\S4method{pbm_condition}{CoRecMotif}(x) <- value

\S4method{zscore_motif}{CoRecMotif}(x) <- value

\S4method{array_id}{CoRecMotif}(x) <- value

\S4method{seed_sequence}{CoRecMotif}(x) <- value

\S4method{motif_name}{CoRecMotif}(x) <- value

\S4method{as.data.frame}{CoRecMotif}(x)

\S4method{show}{CoRecMotif}(object)
}
\description{
A container for motifs generated by a CoRec experiment.
}
\section{Slots}{

\describe{
\item{\code{probe_set}}{\code{character(1)}. The probe set name.}

\item{\code{pbm_condition}}{\code{character(1)}. The PBM condition (e.g., cell type,
treatment, and factor profiled).}

\item{\code{zscore_motif}}{\code{matrix(numeric())}. The z-score motif. See
\code{\link[=make_zscore_motif]{make_zscore_motif()}} for a description of the expected format.}

\item{\code{array_id}}{\code{character(1)}. The name of the array/experiment the z-score
motif is from.}

\item{\code{motif_strength}}{\code{numeric(1)}. Calculated automatically. The motif
strength is the 93rd percentile of the z-scores in the z-score motif (or
approximately the median of the top 15\% of z-scores).}

\item{\code{rolling_ic}}{\code{numeric(1)}. Calculated automatically. The rolling IC is
the maximum value obtained when taking a moving average of the information
content of the motif over a sliding window of length 5.}

\item{\code{seed_sequence}}{\code{character(1)}. The sequence of the seed probe.}

\item{\code{beta}}{\code{numeric(1)}. Calculated automatically. Beta is defined as \eqn{4
  - (0.5 * z)}, where \code{z} is the z-score of the seed probe of this probe set.
If beta falls outside the range of 1 to 4 (inclusive), it is set to the
nearest endpoint. It is used to convert the z-score motif to a PPM.}

\item{\code{motif}}{\link[universalmotif:universalmotif-class]{universalmotif}.
Calculated automatically. To convert the z-score motif into a PPM, each
z-score is first multiplied by beta, and then the product is exponentiated.
Each column (or position in the motif) is then normalized to sum to 1.}

\item{\code{match_motif}}{\link[universalmotif:universalmotif-class]{universalmotif}. The
best match reference motif.}

\item{\code{match_pvalue}}{\code{numeric(1)}. The adjusted p-value of the best match to a
reference motif.}

\item{\code{match_cluster}}{\code{character(1)}. The cluster of the best match reference
motif.}
}}

