% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corecmotif-class.R
\name{CoRecMotif}
\alias{CoRecMotif}
\title{Make a new CoRecMotif}
\usage{
CoRecMotif(
  probe_set,
  pbm_condition,
  zscore_motif,
  array_id,
  seed_sequence = NA_character_
)
}
\arguments{
\item{probe_set}{a string specifying the probe set name.}

\item{pbm_condition}{a string specifying the PBM condition (e.g., cell type,
treatment, and factor profiled).}

\item{zscore_motif}{a 4 x n numeric matrix of z-scores such as generated by
\code{\link[=make_zscore_motif]{make_zscore_motif()}}.}

\item{array_id}{\code{character(1)} or \code{NULL}. The name of the array/experiment
the z-score motif is from.}

\item{seed_sequence}{a string specifying the sequence of the seed probe.}
}
\value{
An object of class \link[=CoRecMotif-class]{CoRecMotif}.
}
\description{
Create a CoRecMotif from a 4 x n matrix of z-scores.
}
\examples{
print("FILL THIS IN")
}
