% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_corecmotifs}
\alias{example_corecmotifs}
\alias{example_matched_corecmotifs}
\title{Example CoRecMotifs}
\format{
A list of 10 or 40 \link[=CoRecMotif-class]{CoRecMotifs}
\subsection{example_corecmotifs}{

A list of 40 \link[=CoRecMotif-class]{CoRecMotifs} that have not been matched to
reference motifs.
}

\subsection{example_matched_corecmotifs}{

A list of 10 \link[=CoRecMotif-class]{CoRecMotifs} that have been matched to
reference motifs.
}
}
\usage{
example_corecmotifs

example_matched_corecmotifs
}
\description{
Lists of example \link[=CoRecMotif-class]{CoRecMotifs}.
}
\keyword{datasets}
