% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fluorescence_table}
\alias{example_fluorescence_table}
\alias{example_annotated_fluorescence_table}
\alias{example_zscore_table}
\title{Example fluorescence and z-score data}
\format{
A data frame with 533 rows and 5 or 10 columns:
\subsection{example_fluorescence_table}{

The un-annotated fluorescence table contains 5 columns: \code{probe_id}, and 4
columns of raw fluorescence data named \code{UT_SUDHL4_SWISNF_mix},
\code{UT_SUDHL4_HDAC_mix}, \code{UT_SUDHL4_PRMT5}, and \code{UT_SUDHL4_JMJD2A}.
}

\subsection{example_annotated_fluorescence_table}{

The annotated fluorescence table contains the 6 probe information columns
described in \link{hTF_v1_annotation} and the 4 raw fluorescence data columns
described above.
}

\subsection{example_zscore_table}{

The z-score table contains the same columns as the annotated fluorescence
table after converting the raw fluorescence data into z-scores.
}
}
\usage{
example_fluorescence_table

example_annotated_fluorescence_table

example_zscore_table
}
\description{
Tables of example CoRec fluorescence data and z-score data for 5 probe sets
in 4 conditions.
}
\keyword{datasets}
