% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_motifs.R
\name{make_corec_motifs}
\alias{make_corec_motifs}
\alias{make_zscore_motif}
\title{Extract corecmotif objects from z-score matrix}
\usage{
make_corec_motifs(zscore_table, pbm_conditions)

make_zscore_motif(zscore_table, seed_name, pbm_condition)
}
\arguments{
\item{zscore_table}{A data frame of PBM condition-wise fluorescence value
z-scores, such as output by fluorescence_to_zscore_table().}

\item{seed_name}{The name of the seed for which to extract a corecmotif
object.}

\item{pbm_condition, pbm_conditions}{A character vector of the column names of
the PBM conditions for which to extract corecmotif objects.}

\item{pbm_condition}{The column name of the PBM condition for which to
extract a corecmotif object.}
}
\value{
A list of corecmotif objects, one for each possible combination of
the seed names present in \code{zscore_table} and the PBM conditions in
\code{pbm_conditions}.

A data frame representing the z-score motif for the seed name and PBM
condition provided to \code{seed_name} and \code{pbm_condition}
respectively, where the rows are nucleotides and the columns are positions
in the motif.
}
\description{
Extracts a corecmotif object from the input z-score matrix for every possible
combination of the given seed names and PBM conditions.
}
\examples{
corec_motifs <-
    zscore_table_to_motifs(
        zscore_table,
        seed_names,
        pbm_conditions_br
    )
zscore_motif <-
    make_zscore_motif(
        zscore_table,
        "MA0079.3_SP1",
        "v1_a11_run1_UT_SUDHL4_SMARCA4MIX"
    )
}
