% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_match.R
\name{find_match}
\alias{find_match}
\title{Match a list of CoRecMotifs to reference motifs}
\usage{
find_match(
  corecmotifs,
  reference_motifs_file,
  cluster_assignments = NULL,
  meme_path = NULL,
  min_overlap = 5,
  output_file = NULL
)
}
\arguments{
\item{corecmotifs}{the list of \link[=CoRecMotif-class]{CoRecMotifs} to match to
reference motifs.}

\item{reference_motifs_file}{the path to a MEME format file of reference
motifs to match to.}

\item{cluster_assignments}{a data frame mapping the reference motifs to motif
clusters or NULL to skip the cluster assignment step.}

\item{meme_path}{the path to "meme/bin/" or NULL to rely on
\code{\link[memes:runTomTom]{memes::runTomTom()}} to find it.}

\item{min_overlap}{a single positive integer specifying the minimum amount of
overlap to require when comparing a \link[=CoRecMotif-class]{CoRecMotif} to a
reference motif.}

\item{output_file}{the path to the RDS file where the list of matched
\link[=CoRecMotif-class]{CoRecMotifs} will be written. If NULL (the default), no
file is written.}
}
\value{
A list of \link[=CoRecMotif-class]{CoRecMotifs} with the match_motif,
match_pvalue, and (optionally) match_cluster slots filled in.
}
\description{
Identifies the reference motif that is the best match to each
\link[=CoRecMotif-class]{CoRecMotif}.
}
\details{
The PPM representation of each \link[=CoRecMotif-class]{CoRecMotif} is compared to
the reference motifs with \code{\link[memes:runTomTom]{memes::runTomTom()}} using Euclidean distance
("ed") as the distance metric. The match_motif and match_pvalue slots are
filled in based on the best match returned by \code{\link[memes:runTomTom]{memes::runTomTom()}}. The
match_pvalue slot is corrected for multiple comparisons by multiplying the
raw match p-value by the number of motifs in the reference database. The
match_cluster slot is filled in based on the user-provided cluster
assignments or left empty if no cluster assignments are provided.
}
\examples{
print("FILL THIS IN")
}
