% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_matrix.R
\name{fluorescence_to_zscore_matrix}
\alias{fluorescence_to_zscore_matrix}
\title{Transform fluorescence matrix to z-scores}
\usage{
fluorescence_to_zscore_matrix(annotated_fluorescence_matrix)
}
\arguments{
\item{annotated_fluorescence_matrix}{A data frame of fluorescence values and
annotations for each probe, such as output by
annotate_fluorescence_matrix().}
}
\value{
A data frame of PBM condition-wise fluorescence value z-scores.
}
\description{
Converts a data frame of fluorescence values into PBM condition-wise z-scores
using the background fluorescence values.
}
\details{
This function normalizes raw fluorescence values against the fluorescence
values of the background probes. It first log-transforms all the fluorescence
values for a given PBM condition, then calculates a z-score like statistic
using the mean and standard deviation of the background probes.
}
\examples{
zscore_matrix <- fluorescence_to_zscore_matrix(annotated_fluorescence_matrix)
}
