% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_fluorescence_to_zscore_table.R
\name{fluorescence_to_zscore_table}
\alias{fluorescence_to_zscore_table}
\title{Transform fluorescence values into z-scores}
\usage{
fluorescence_to_zscore_table(
  fluorescence_table,
  fluorescence_columns,
  output_file = NULL
)
}
\arguments{
\item{fluorescence_table}{\code{data.frame}. An annotated table of fluorescence
values. See \link{hTF_v1_annotation} for expected annotation columns.}

\item{fluorescence_columns}{\code{character}. The names of the columns of
\code{fluorescence_table} that contain fluorescence data.}

\item{output_file}{\code{character(1)} or \code{NULL}. The path to the file where the
results should be saved, or NULL not to save the results. (Default: NULL)}
}
\value{
A data frame of fluorescence z-scores and the corresponding probe
information. See \link{hTF_v1_annotation} for a description of the probe
annotation columns.
}
\description{
Converts an annotated data frame of fluorescence data into z-scores.
}
\details{
Raw fluorescence values are normalized against the fluorescence values of the
background probes. Specifically, all the fluorescence values for a given PBM
condition are log-transformed. Then a z-score-like statistic is calculated
using the mean and standard deviation of the fluorescence values of the
background probes in that condition.
}
\examples{
print("FILL THIS IN")
}
\seealso{
\link{hTF_v1_annotation} for a description of the probe annotation
columns.
}
