% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_comparisons.R
\name{identify_motif_match}
\alias{identify_motif_match}
\title{Title}
\usage{
identify_motif_match(corec_motif, reference_motifs, score_threshold = 1.5, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Identifies the reference motif that is the best match to the corecmotif and
calculates a p-value for that match.
}
