% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_fluorescence_data}
\alias{load_fluorescence_data}
\title{Import fluorescence data}
\usage{
load_fluorescence_data(fluorescence_file, pbm_conditions)
}
\arguments{
\item{fluorescence_file}{\code{character(1)}. The path to the fluorescence data
file to load. See 'Details' for expected columns.}

\item{pbm_conditions}{\code{character}. The PBM conditions (e.g., cell type,
treatment, and factor profiled) in the order they appear in
\code{fluorescence_file}.}
}
\value{
A data frame containing the fluorescence values from the specified
fluorescence data file. The first column will be named "probe_id", and the
remaining columns will be named according to the PBM conditions supplied to
\code{pbm_conditions}.
}
\description{
Reads a file containing fluorescence data (in which rows correspond to probes
and columns correspond to fluorescence values) and reformats it. The output
from this function is suitable for use by \link{annotate_fluorescence_table}.
}
\details{
The fluorescence data file supplied to \code{fluorescence_file} must contain
\eqn{n + 3} tab-delimited columns, where n is the number of PBM conditions
supplied to \code{pbm_conditions}. This file must NOT have a header. The columns
must be as follows:
\enumerate{
\item The probe ID. The orientation tags (i.e., "_o1" or "_o2") will be removed
from these IDs.
\item The probe sequence. This column will be removed.
\item The number of PBM conditions profiled on this array. This column will be
removed.
}

Columns 4 through \eqn{(n + 3)} must contain the fluorescence values for the
PBM conditions in the order they appear in \code{pbm_conditions}.
}
\examples{
print("FILL THIS IN")
}
