% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_motifs.R
\name{make_corec_motifs}
\alias{make_corec_motifs}
\title{Create \linkS4class{corecmotif} objects from a table of z-score data}
\usage{
make_corec_motifs(zscore_table, zscore_columns, output_file = NULL)
}
\arguments{
\item{zscore_table}{a data frame of z-scores and annotations for each probe.
See 'Details' of \code{\link{make_fluorescence_table}} for a description of
the expected annotation columns.}

\item{zscore_columns}{a character vector specifying the names of the columns
of \code{zscore_table} that contain z-score data.}

\item{output_file}{the name of the RDS file where the list of corecmotif
objects will be written. If NULL (the default), no file is written.}
}
\value{
A list of \linkS4class{corecmotif} objects, one for each possible
combination of the probe sets in \code{zscore_table} and the PBM conditions
listed in \code{zscore_columns}.
}
\description{
Creates a list of \linkS4class{corecmotif} objects for all possible
combinations of the probe sets present in \code{zscore_table} and the PBM
conditions given in \code{zscore_columns}.
}
\examples{
# Load the example z-score table
zscore_table <-
    read.table(
        "example_data/example_output/example_v1_a11_run1_zscores.tsv",
        header = TRUE,
        sep = "\t",
        stringsAsFactors = FALSE
    )

# Make a list of corecmotif objects
corec_motifs <-
    make_corec_motifs(
        zscore_table,
        zscore_columns = c(
            "v1_a11_run1_UT_SUDHL4_SMARCA4MIX",
            "v1_a11_run1_UT_SUDHL4_HDAC1MIX",
            "v1_a11_run1_UT_SUDHL4_SUZ12",
            "v1_a11_run1_UT_SUDHL4_PRMT5"
        )
    )
}
