% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_run_full_analysis.R
\name{make_corecmotifs}
\alias{make_corecmotifs}
\title{Make a list of CoRecMotifs}
\usage{
make_corecmotifs(
  fluorescence_table,
  fluorescence_columns,
  annotation,
  array_id = NULL,
  output_directory = NULL,
  output_base_name = NULL
)
}
\arguments{
\item{fluorescence_table}{\code{data.frame}. A table of fluorescence values.}

\item{fluorescence_columns}{\code{character}. The names of the columns of
\code{fluorescence_table} that contain fluorescence data.}

\item{annotation}{\code{data.frame}. The probe annotations to add to the
fluorescence table. See \link{hTF_v1_annotation} for expected annotation
columns.}

\item{array_id}{\code{character(1)} or \code{NULL}. The name of the array/experiment
the fluorescence z-scores are from. If \code{NULL}, a random ID will be
generated. (Default: NULL)}

\item{output_directory}{\code{character(1)} or \code{NULL}. The path to the directory
where output files will be saved. No output files will be created unless
\code{output_directory} or \code{output_base_name} is provided. (Default: NULL)}

\item{output_base_name}{\code{character(1)} or \code{NULL}. The base name for output
files. No output files will be created unless \code{output_directory} or
\code{output_base_name } is provided. (Default: NULL)}
}
\value{
A list of \link[=CoRecMotif-class]{CoRecMotifs}, one for each possible
combination of the probe sets annotated in \code{annotation} and the PBM
conditions listed in \code{pbm_conditions}.
}
\description{
Creates a list of \link[=CoRecMotif-class]{CoRecMotifs} using fluorescence data
from a CoRec experiment. This is a convenience function that calls
\code{\link[=annotate_fluorescence_table]{annotate_fluorescence_table()}}, \code{\link[=fluorescence_to_zscore_table]{fluorescence_to_zscore_table()}}, and then
\code{\link[=zscore_table_to_corecmotifs]{zscore_table_to_corecmotifs()}}.
}
\details{
By default no output files are created. To save output files, you must
provide \code{output_directory}, \code{output_base_name}, or both. The output files are
named according to the following rules.
\itemize{
\item If \code{output_directory} is provided but \code{output_base_name} is not, the files
will be saved in the provided directory using the base name "output".
\item If \code{output_base_name} is provided but \code{output_directory} is not, the files
will be saved in the working directory using the provided base name.
\item If \code{array_id} is provided, it will be appended to the output base name.
\item The individual output file names will be identified by a suffix appended
to the output base name after the array ID (if present).
\item The output of \code{\link[=annotate_fluorescence_table]{annotate_fluorescence_table()}} will have the suffix
"fluorescence.tsv".
\item The output of \code{\link[=fluorescence_to_zscore_table]{fluorescence_to_zscore_table()}} will have the suffix
"zscores.tsv"
\item The output of \code{\link[=zscore_table_to_corecmotifs]{zscore_table_to_corecmotifs()}} will have the suffix
"all_corecmotifs.rds"
}
}
\examples{
print("FILL THIS IN!")
}
\seealso{
\code{\link[=annotate_fluorescence_table]{annotate_fluorescence_table()}}, \code{\link[=fluorescence_to_zscore_table]{fluorescence_to_zscore_table()}},
\code{\link[=zscore_table_to_corecmotifs]{zscore_table_to_corecmotifs()}}
}
