% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_table_to_corecmotifs.R
\name{make_zscore_motif}
\alias{make_zscore_motif}
\title{Create a z-score motif for a given probe set and PBM condition}
\usage{
make_zscore_motif(zscore_table, probe_set, pbm_condition)
}
\arguments{
\item{zscore_table}{a data frame of z-scores and annotations for each probe.
See 'Details' of \code{\link{annotate_fluorescence_table}} for a
description of the expected annotation columns.}

\item{probe_set}{a character string containing the name of the probe set
for which to create the z-score motif.}

\item{pbm_condition}{a character string containing the name of the PBM
condition for which to create the z-score motif.}
}
\value{
A data frame with rows corresponding to nucleotides and columns
corresponding to positions in the motif. Each cell is filled with the
z-score of the relevant probe in the given PBM condition.
}
\description{
Creates a z-score motif for a given probe set and PBM condition. The rows
correspond to nucleotides and the columns correspond to positions in the
motif.
}
\examples{
# Load the example z-score table
zscore_table <-
    read.table(
        "example_data/output/example_rep1_v1_a11_run1_zscores.tsv",
        header = TRUE,
        sep = "\t",
        stringsAsFactors = FALSE
    )

# Create a z-score motif
zscore_motif <-
    make_zscore_motif(
        zscore_table,
        probe_set = "MA0052.3_MEF2A",
        pbm_condition = "UT_SUDHL4_PRMT5"
    )
}
