% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_zscore_table.R
\name{make_zscore_table}
\alias{make_zscore_table}
\title{Transform fluorescence values into z-scores}
\usage{
make_zscore_table(fluorescence_table, fluorescence_columns)
}
\arguments{
\item{fluorescence_table}{a data frame of fluorescence values and annotations
for each probe. See the return value of
\code{\link{make_fluorescence_table}} for a description of the expected
columns.}

\item{fluorescence_columns}{a character vector specifying the names of the
columns of \code{fluorescence_table} that contain fluorescence data.}
}
\value{
A data frame of column-wise fluorescence z-scores and probe
annotations.
}
\description{
Converts an annotated data frame of fluorescence data into z-scores.
}
\details{
Raw fluorescence values are normalized against the fluorescence values of the
background probes. Specifically, all the fluorescence values for a given PBM
condition are log-transformed. Then a z-score-like statistic is calculated
using the mean and standard deviation of the fluorescence values of the
background probes in that condition.
}
\examples{
fluorescence_table <- make_fluorescence_table(
    fluorescence_file =
        "example_data/data_matrices/hTF_v1_SUDHL4_14jan21_br.dat",
    pbm_conditions = c(
        "UT_SUDHL4_SMARCA4MIX",
        "UT_SUDHL4_HDAC1MIX",
        "UT_SUDHL4_SUZ12",
        "UT_SUDHL4_PRMT5",
        "UT_SUDHL4_JMJD2A",
        "UT_SUDHL4_BMI1",
        "UT_SUDHL4_DNMT3A",
        "UT_SUDHL4_ASH2L"
    ),
    annotation_file = "example_data/hTF_v01_PBM_ANNOT.txt",
    array_id = "v1_a6_run1"
)

zscore_table <- make_zscore_table(
    fluorescence_table = fluorescence_table,
    fluorescence_columns = c(
        "v1_a11_run1_UT_SUDHL4_SMARCA4MIX",
        "v1_a11_run1_UT_SUDHL4_HDAC1MIX",
        "v1_a11_run1_UT_SUDHL4_SUZ12",
        "v1_a11_run1_UT_SUDHL4_PRMT5",
        "v1_a11_run1_UT_SUDHL4_JMJD2A",
        "v1_a11_run1_UT_SUDHL4_BMI1",
        "v1_a11_run1_UT_SUDHL4_DNMT3A",
        "v1_a11_run1_UT_SUDHL4_ASH2L"
    )
)
}
