% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_clusters}
\alias{motif_clusters}
\title{Motif cluster assignments}
\format{
A data frame with 946 rows and 2 columns:
\describe{
\item{motif}{The name of a reference motif.}
\item{cluster}{The name of the cluster this reference motif is part of.}
}
}
\usage{
motif_clusters
}
\description{
A table assigning reference motifs to clusters.
}
\keyword{datasets}
