% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_plots.R
\name{plot_corecmotif}
\alias{plot_corecmotif}
\title{Title}
\usage{
plot_corecmotif(
  corec_motif,
  motif_logo_type = c("delta_zscore", "PPM", "PWM", "ICM"),
  seed_zscore_threshold = 1,
  rolling_ic_threshold = 1.5,
  outline_color = "#abbed1"
)
}
\arguments{
\item{motif_score_threshold}{}
}
\value{

}
\description{
Title
}
