% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_run_full_analysis.R
\name{process_corecmotifs}
\alias{process_corecmotifs}
\title{Filter a list of CoRecMotifs and match them to reference motifs}
\usage{
process_corecmotifs(
  corecmotifs,
  reference_motifs_file,
  cluster_assignments = NULL,
  meme_path = NULL,
  motif_strength = 1,
  rolling_ic = 1,
  n_replicates = 2,
  eucl_distance = 0.4,
  min_overlap = 5,
  match_pvalue = 0.01,
  pbm_condition_groups = NULL,
  output_directory = NULL,
  output_base_name = NULL
)
}
\arguments{
\item{corecmotifs}{\code{list}. The \link[=CoRecMotif-class]{CoRecMotifs} to process.}

\item{reference_motifs_file}{\code{character(1)}. The path to the MEME format file
of reference motifs to match to.}

\item{cluster_assignments}{\code{data.frame} or \code{NULL}. A table mapping the
reference motifs to motif clusters or NULL to skip the cluster assignment
step. See \link{motif_clusters} for expected columns. (Default: NULL)}

\item{meme_path}{\code{character(1)} or \code{NULL}. The path to "meme/bin/" or NULL to
rely on \code{\link[memes:runTomTom]{memes::runTomTom()}} to find it. (Default: NULL)}

\item{motif_strength}{\code{numeric(1)} or \code{NULL}. The minimum motif strength to
keep or NULL not to filter by motif strength. (Default: 1)}

\item{rolling_ic}{\code{numeric(1)} or \code{NULL}. The minimum rolling IC to keep or
NULL not to filter by rolling IC. (Default: 1)}

\item{n_replicates}{\code{integer(1)}. The minimum number of replicates to
require. Set this to 1 to skip filtering by number of replicates. (Default:
2)}

\item{eucl_distance}{\code{numeric(1)} or \code{NULL}. The maximum allowable Euclidean
distance between replicate motifs or NULL to skip the replicate comparison
step. (Default: 0.4)}

\item{min_overlap}{\code{integer(1)}. The minimum amount of overlap to require
when comparing a \link[=CoRecMotif-class]{CoRecMotif} to a reference motif.
(Default: 5)}

\item{match_pvalue}{\code{numeric(1)} or \code{NULL}. The maximum match p-value to keep
or NULL not to filter by match p-value. (Default: 0.01)}

\item{pbm_condition_groups}{\code{list(character)} or \code{NULL}. The names of the PBM
conditions to compare. Each element of the list should contain a group of
PBM conditions to compare to each other. If the list elements are named,
the names will be passed to the \code{group_name} parameter of
\code{\link[=compare_conditions]{compare_conditions()}}. (Default: NULL)}

\item{output_directory}{\code{character(1)} or \code{NULL}. The path to the directory
where output files will be saved. No output files will be created unless
\code{output_directory} or \code{output_base_name} is provided. (Default: NULL)}

\item{output_base_name}{\code{character(1)} or \code{NULL}. The base name for output
files. No output files will be created unless \code{output_directory} or
\code{output_base_name } is provided. (Default: NULL)}
}
\value{
A filtered list of replicated \link[=CoRecMotif-class]{CoRecMotifs} that
match a reference motif.
}
\description{
Removes \link[=CoRecMotif-class]{CoRecMotifs} with low z-scores or motif strengths
and CoRecMotifs that do not replicate, then compares the remaining
CoRecMotifs to a database of reference motifs to identify the best match.
This is a convenience function that calls \code{\link[=filter_corecmotifs]{filter_corecmotifs()}},
\code{\link[=check_replicates]{check_replicates()}}, \code{\link[=find_match]{find_match()}}, then \code{\link[=filter_corecmotifs]{filter_corecmotifs()}} and
\code{\link[=check_replicates]{check_replicates()}} again, and finally \code{\link[=summarize_corecmotifs]{summarize_corecmotifs()}} and
optionally \code{\link[=compare_conditions]{compare_conditions()}}.
}
\details{
By default no output files are created. To save output files, you must
provide \code{output_directory}, \code{output_base_name}, or both. The output files are
named according to the following rules.
\itemize{
\item If \code{output_directory} is provided but \code{output_base_name} is not, the files
will be saved in the provided directory using the base name "output".
\item If \code{output_base_name} is provided but \code{output_directory} is not, the files
will be saved in the working directory using the provided base name.
\item The individual output file names will be identified by a suffix appended
to the output base name.
\item The output after the first call to \code{\link[=filter_corecmotifs]{filter_corecmotifs()}} and
\code{\link[=check_replicates]{check_replicates()}} will have the suffix "filtered_corecmotifs.rds".
\item The output of \code{\link[=find_match]{find_match()}} will have the suffix
"matched_corecmotifs.rds".
\item The output after the second call to \code{\link[=filter_corecmotifs]{filter_corecmotifs()}} and
\code{\link[=check_replicates]{check_replicates()}} will have the suffix "significant_corecmotifs.rds".
\item The output of \code{\link[=summarize_corecmotifs]{summarize_corecmotifs()}} will have the suffix
"significant_corecmotifs_summary.tsv".
\item The output of \code{\link[=summarize_corecmotifs]{summarize_corecmotifs()}} with \code{by_cluster = TRUE} will have
the suffix "significant_corecmotifs_summary_by_cluster.tsv".
\item The output of \code{\link[=compare_conditions]{compare_conditions()}} will have the suffix
"condition_comparisons.tsv".
}
}
\examples{
print("FILL THIS IN")
}
\seealso{
\code{\link[=filter_corecmotifs]{filter_corecmotifs()}}, \code{\link[=check_replicates]{check_replicates()}}, \code{\link[=find_match]{find_match()}},
\code{\link[=summarize_corecmotifs]{summarize_corecmotifs()}}, \code{\link[=compare_conditions]{compare_conditions()}}
}
