% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_matrix.R
\name{read_fluorescence_matrices}
\alias{read_fluorescence_matrices}
\title{Import fluorescence matrix}
\usage{
read_fluorescence_matrices(
  matrix_directory,
  matrix_base_name,
  pbm_conditions,
  run_tag = NA
)
}
\arguments{
\item{matrix_directory}{The path to the directory containing the individual
o1, o2, or, and br fluorescence matrices to import.}

\item{matrix_base_name}{The base filename of the fluorescence matrices to
import.}

\item{pbm_conditions}{A character vector of the names of the PBM conditions.
These will be used as column names.}

\item{run_tag}{An optional (but recommended) tag specifying the particular
experiment these fluorescence matrices are from. It will be incorporated
into the column names along with the PBM conditions. This option is useful
for differentiating replicates when merging two matrices downstream.}
}
\value{
A data frame containing the fluorescence values from the specified
set of o1, o2, or, and br matrices.
}
\description{
Reads a set of individual o1, o2, or, and br fluorescence matrices and
creates a single data frame from them, with rows corresponding to probes and
columns corresponding to PBM conditions.
}
\details{
This function takes as input the fluorescence matrices output by the PBM
preprocessing pipeline. It expects the filenames to be of the form
\code{<matrix_base_name>_<probe_orientation>.dat}, where
\code{<matrix_base_name>} is the value passed to the \code{matrix_base_name}
parameter and \code{<probe_orientation>} is \verb{"o1"}, \verb{"o2"},
\verb{"or"}, or \verb{"br"} (all four files must be present). These files
should contain \eqn{3 + n} tab-delimited columns, where n is the number of
PBM conditions supplied to the \code{pbm_conditions} parameter. The first
column must contain the probe IDs, and the fourth through nth columns must
contain the fluorescence values for the PBM conditions in the order listed.
The second and third columns are removed. The files should not contain
headers.

The output data frame contains \eqn{1 + (4 * n)} columns, where n is again
the number of PBM conditions supplied to the \code{pbm_conditions} parameter.
The first column will be named \verb{"probeID"} and will contain the probe
IDs with their orientation tags (i.e., \verb{"_o1"} or \verb{"_o2"}) removed.
The remaining columns will contain the fluorescence values for each probe
orientation and PBM condition and will be named based on the vector of PBM
conditions supplied to \code{pbm_conditions} and the tag supplied to
\code{run_tag} (if any). If a tag is provided, the column names will be of
the form \verb{"<run_tag>_<probe_orientation>_<pbm_condition>"}. If no tag is
provided, the column names will be of the form
\verb{"<probe_orientation>_<pbm_condition>"}.
}
\examples{
fluorescence_matrix <- read_fluorescence_matrices(
    matrix_directory = paste0(
        "/projectnb/siggers/data/hTF_array_project/data/data_matrices/",
        "v1_a11_run1"
    ),
    matrix_base_name = "hTF_v1_SUDHL4_14jan21",
    pbm_conditions = c(
        "UT_SUDHL4_SMARCA4MIX",
        "UT_SUDHL4_HDAC1MIX",
        "UT_SUDHL4_SUZ12",
        "UT_SUDHL4_PRMT5",
        "UT_SUDHL4_JMJD2A",
        "UT_SUDHL4_BMI1",
        "UT_SUDHL4_DNMT3A",
        "UT_SUDHL4_ASH2L"
    )
)
}
