% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pipeline.R
\name{run_full_analysis}
\alias{run_full_analysis}
\title{Title}
\usage{
run_full_analysis(
  output_directory,
  matrix_directory,
  matrix_base_name,
  pbm_conditions_file = NA,
  pbm_conditions = NA,
  annotation_file,
  reference_motifs_file,
  run_tag = NA,
  seed_zscore_threshold = 1,
  rolling_ic_threshold = 1.5,
  comparison_method = "ed",
  pvalue_threshold = 0.05
)
}
\arguments{
\item{run_tag}{}
}
\value{

}
\description{
Title
}
