% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_scores.R
\name{score_motif}
\alias{score_motif}
\title{Score a corecmotif}
\usage{
score_motif(
  corec_motif,
  method = c("seed_zscore", "mean_highest_delta_zscore", "median_highest_delta_zscore",
    "rolling_mean_highest_delta_zscore", "rolling_median_highest_delta_zscore",
    "mean_information_content", "median_information_content",
    "rolling_mean_information_content", "rolling_median_information_content"),
  window_length = 5
)
}
\arguments{
\item{corec_motif}{The corecmotif object to score.}

\item{method}{The scoring metric to use. Must be one of ..............}

\item{window_length}{The length of the sliding window to use for metrics
involving rolling means or medians.}
}
\value{
A corecmotif object with the motif_score and motif_score_type slots
filled in. If the corecmotif object supplied to the \code{corec_motif}
argument already has values in these slots, they will be overwritten.
}
\description{
Assign a score to a corecmotif object using one of several available scoring
metrics. See 'Details' for descriptions of the available metrics.
}
\examples{
scored_corec_motif <- score_motif(corec_motif)
}
