% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corecmotif-class.R
\name{update_cluster_match}
\alias{update_cluster_match}
\title{Update the motif cluster match of a CoRecMotif}
\usage{
update_cluster_match(corecmotif, cluster_assignments = NULL)
}
\arguments{
\item{cluster_assignments}{}
}
\value{

}
\description{
Update the motif cluster match of a CoRecMotif
}
