% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_annotate_fluorescence_table.R
\name{annotate_fluorescence_table}
\alias{annotate_fluorescence_table}
\title{Annotate fluorescence data}
\usage{
annotate_fluorescence_table(
  fluorescence_table,
  fluorescence_columns,
  annotation,
  output_file = NULL
)
}
\arguments{
\item{fluorescence_table}{\code{data.frame}. A table of fluorescence values.}

\item{fluorescence_columns}{\code{character}. The names of the columns of
\code{fluorescence_table} that contain fluorescence data.}

\item{annotation}{\code{data.frame}. The probe annotations to add to the
fluorescence table. See \link{hTF_v1_annotation} for expected annotation
columns.}

\item{output_file}{\code{character(1)} or \code{NULL}. The path to the file where the
results should be saved, or NULL not to save the results. (Default: NULL)}
}
\value{
A data frame of fluorescence values and the corresponding probe
information. See \link{hTF_v1_annotation} for a description of the probe
annotation columns.
}
\description{
Adds probe information to a table of fluorescence data from a CoRec
experiment.
}
\details{
The fluorescence data table must have a column named \code{probe_id}. The
remaining column names must match the names listed in \code{fluorescence_columns}.
Any additional columns will be dropped.

See \link{hTF_v1_annotation} for a description of the expected columns of the
annotation table. Any additional columns will be dropped.

The IDs in the \code{probe_id} column in the annotation table must match the IDs
in the \code{probe_id} column of the fluorescence table. IDs that are present in
the annotation table but not the fluorescence table will be dropped silently.
IDs that are present in the fluorescence table but not the annotation table
will be dropped with a warning.
}
\examples{
print("FILL THIS IN")
}
\seealso{
\link{hTF_v1_annotation} for a description of the probe annotation
columns.
}
