% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_conditions.R
\name{compare_conditions}
\alias{compare_conditions}
\title{Compare CoRecMotifs across different PBM conditions}
\usage{
compare_conditions(
  corecmotifs,
  pbm_conditions,
  pbm_conditions_group = NULL,
  eucl_distance = 0.25,
  check_corecmotifs = TRUE
)
}
\arguments{
\item{corecmotifs}{\code{list}. The \link[=CoRecMotif-class]{CoRecMotifs} to compare to
each other.}

\item{pbm_conditions}{\code{character}. The names of the individual PBM conditions
to compare.}

\item{pbm_conditions_group}{\code{character(1)} or \code{NULL}. The name of the group
of PBM conditions to compare. This name will be combined with the probe set
ID and a number to identify groups of motifs that are similar across
conditions. (Default: NULL)}

\item{eucl_distance}{\code{numeric(1)}. The maximum allowable Euclidean distance
between conditions to group. (Default: 0.4)}

\item{check_corecmotifs}{\code{logical(1)}. Should \code{corecmotifs} be checked for
validity? Setting this to FALSE can increase speed, but if \code{corecmotifs}
contains anything other than valid \link[=CoRecMotif-class]{CoRecMotifs}, it may
produce uninformative error messages. (Default: TRUE)}
}
\value{
A data frame with comparison information about a list of
\link[=CoRecMotif-class]{CoRecMotifs}.
}
\description{
Compares the \link[=CoRecMotif-class]{CoRecMotifs} in each PBM condition to the
CoRecMotifs from the same probe set in all the other PBM conditions.
}
\examples{
print("FILL THIS IN")
}
