% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRecMotif-class.R, R/CoRecMotif-methods.R
\docType{class}
\name{CoRecMotif-class}
\alias{CoRecMotif-class}
\alias{as.data.frame,CoRecMotif-method}
\alias{show,CoRecMotif-method}
\title{CoRecMotif class}
\usage{
\S4method{as.data.frame}{CoRecMotif}(x)

\S4method{show}{CoRecMotif}(object)
}
\arguments{
\item{x, object}{\link[=CoRecMotif-class]{CoRecMotif}. The motif to manipulate.}
}
\description{
A container for motifs generated by a CoRec experiment.
}
\section{Slots}{

\describe{
\item{\code{probe_set}}{\code{character(1)}. The probe set name.}

\item{\code{pbm_condition}}{\code{character(1)}. The PBM condition (e.g., cell type,
treatment, and factor profiled).}

\item{\code{zscore_motif}}{\code{matrix(numeric())}. The z-score motif. See
\code{\link[=make_zscore_motif]{make_zscore_motif()}} for a description of the expected format.}

\item{\code{array_id}}{\code{character(1)}. The name of the array/experiment the z-score
motif is from.}

\item{\code{motif_strength}}{\code{numeric(1)}. Calculated automatically. The motif
strength is the 93rd percentile of the z-scores in the z-score motif (or
approximately the median of the top 15\% of z-scores).}

\item{\code{rolling_ic}}{\code{numeric(1)}. Calculated automatically. The rolling IC is
the maximum value obtained when taking a moving average of the information
content of the motif over a sliding window of length 5.}

\item{\code{seed_sequence}}{\code{character(1)}. The sequence of the seed probe.}

\item{\code{beta}}{\code{numeric(1)}. Calculated automatically. Beta is defined as \eqn{4
  - (0.5 * z)}, where \code{z} is the z-score of the seed probe of this probe set.
If beta falls outside the range of 1 to 4 (inclusive), it is set to the
nearest endpoint. It is used to convert the z-score motif to a PPM.}

\item{\code{motif}}{\link[universalmotif:universalmotif-class]{universalmotif}.
Calculated automatically. To convert the z-score motif into a PPM, each
z-score is first multiplied by beta, and then the product is exponentiated.
Each column (or position in the motif) is then normalized to sum to 1.}

\item{\code{match_motif}}{\link[universalmotif:universalmotif-class]{universalmotif}. The
best match reference motif.}

\item{\code{match_pvalue}}{\code{numeric(1)}. The adjusted p-value of the best match to a
reference motif.}

\item{\code{match_cluster}}{\code{character(1)}. The cluster of the best match reference
motif.}
}}

