% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corecmotif-class.R
\docType{class}
\name{CoRecMotif-class}
\alias{CoRecMotif-class}
\title{CoRecMotif class}
\value{
An object of class \code{CoRecMotif}.
}
\description{
A container for motifs generated by a CoRec experiment.
}
\section{Slots}{

\describe{
\item{\code{probe_set}}{a string specifying the probe set name.}

\item{\code{pbm_condition}}{a string specifying the PBM condition (e.g., cell type,
treatment, and factor profiled).}

\item{\code{zscore_motif}}{a 4 x n numeric matrix of z-scores such as generated by
\code{\link[=make_zscore_motif]{make_zscore_motif()}}.}

\item{\code{array_id}}{a string specifying the particular array/experiment the
z-score motif is from.}

\item{\code{motif_strength}}{a number representing the motif strength. Calculated
automatically. See 'Details' for more information}

\item{\code{rolling_ic}}{a number representing the maximum average information
content over a sliding window. Calculated automatically. See 'Details' for
more information.}

\item{\code{seed_sequence}}{a string specifying the sequence of the seed probe.}

\item{\code{beta}}{a number used to convert the z-score motif to a PPM. Calculated
automatically. See 'Details' for more information.}

\item{\code{ppm}}{a \link[universalmotif:universalmotif-class]{universalmotif}
representing the PPM generated using the z-score motif. Calculated
automatically. See 'Details' for more information.}

\item{\code{match_motif}}{a \link[universalmotif:universalmotif-class]{universalmotif}
representing the best match reference motif or NULL if no match has been
found.}

\item{\code{match_pvalue}}{a number representing the adjusted p-value of the best
match to a reference motif.}

\item{\code{match_cluster}}{a string representing the cluster of the best match
reference motif.}
}}

