% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRecMotif-class.R
\name{CoRecMotif}
\alias{CoRecMotif}
\title{Create a CoRecMotif object}
\usage{
CoRecMotif(
  probe_set,
  pbm_condition,
  zscore_motif,
  array_id,
  seed_sequence = NA_character_
)
}
\arguments{
\item{probe_set}{\code{character(1)}. The probe set name.}

\item{pbm_condition}{\code{character(1)}. The PBM condition (e.g., cell type,
treatment, and factor profiled).}

\item{zscore_motif}{\code{matrix(numeric())}. The z-score motif. See
\code{\link[=make_zscore_motif]{make_zscore_motif()}} for a description of the expected format.}

\item{array_id}{\code{character(1)}. The name of the array/experiment the z-score
motif is from.}

\item{seed_sequence}{\code{character(1)}. The sequence of the seed probe.
(Default: NA)}
}
\value{
An object of class \link[=CoRecMotif-class]{CoRecMotif}.
}
\description{
Create a CoRecMotif from a 4 x n matrix of z-scores.
}
\examples{
print("FILL THIS IN")
}
