% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corecmotif-class.R
\name{corecmotif}
\alias{corecmotif}
\title{Title}
\usage{
corecmotif(
  seed_name,
  pbm_condition,
  zscore_motif,
  ic_window_width = 5,
  top_n_percent = 15,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
