% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fluorescence_table_1}
\alias{example_fluorescence_table_1}
\alias{example_fluorescence_table_2}
\title{Example fluorescence data}
\format{
A data frame with 533 rows and 5 columns:
\describe{
\item{probe_id}{The probe ID.}
\item{UT_SUDHL4_SWISNF_mix}{The fluorescence values for the
"UT_SUDHL4_SWISNF_mix" PBM condition.}
\item{UT_SUDHL4_HDAC_mix}{The fluorescence values for the
"UT_SUDHL4_HDAC_mix" PBM condition.}
\item{UT_SUDHL4_PRMT5}{The fluorescence values for the
"UT_SUDHL4_PRMT5" PBM condition.}
\item{UT_SUDHL4_JMJD2A}{The fluorescence values for the
"UT_SUDHL4_SWISNF_mix" PBM condition.}
}
}
\usage{
example_fluorescence_table_1

example_fluorescence_table_2
}
\description{
Tables of example CoRec fluorescence data for 5 probe sets in 4 conditions.
}
\keyword{datasets}
