% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_motifs.R
\name{zscore_matrix_to_motifs}
\alias{zscore_matrix_to_motifs}
\alias{make_zscore_motif}
\title{Extract corecmotif objects from z-score matrix}
\usage{
zscore_matrix_to_motifs(zscore_matrix, seed_names, pbm_conditions)

make_zscore_motif(zscore_matrix, seed_name, pbm_condition)
}
\arguments{
\item{zscore_matrix}{A data frame of PBM condition-wise fluorescence value
z-scores, such as output by fluorescence_to_zscore_matrix().}

\item{seed_names, seed_names}{A character vector of the names of the seeds for
which to extract corecmotif objects.}

\item{seed_name}{The name of the seed for which to extract a corecmotif
object.}

\item{pbm_condition, pbm_conditions}{A character vector of the column names of
the PBM conditions for which to extract corecmotif objects.}

\item{pbm_condition}{The column name of the PBM condition for which to
extract a corecmotif object.}
}
\value{
A list of corecmotif objects, one for each possible combination of
the seed names and PBM conditions provided to \code{seed_names} and
\code{pbm_conditions} respectively.

A data frame representing the z-score motif for the seed name and PBM
condition provided to \code{seed_name} and \code{pbm_condition}
respectively, where the rows are nucleotides and the columns are positions
in the motif.
}
\description{
Extracts a corecmotif object from the input z-score matrix for every possible
combination of the given seed names and PBM conditions.
}
\examples{
corec_motifs <-
    zscore_matrix_to_motifs(
        zscore_matrix,
        seed_names,
        pbm_conditions_br
    )
zscore_motif <-
    make_zscore_motif(
        zscore_matrix,
        "MA0079.3_SP1",
        "v1_a11_run1_br_UT_SUDHL4_SMARCA4MIX"
    )
}
