% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_corecmotifs.R
\name{filter_corecmotifs}
\alias{filter_corecmotifs}
\title{Filter a list of \linkS4class{CoRecMotif} objects}
\usage{
filter_corecmotifs(
  corecmotifs,
  seed_name = NULL,
  pbm_condition = NULL,
  array_id = NULL,
  rolling_ic = NULL,
  motif_strength = NULL,
  seed_sequence = NULL,
  motif_name = NULL,
  match_name = NULL,
  match_altname = NULL,
  match_pvalue = NULL,
  match_cluster = NULL
)
}
\arguments{
\item{corecmotifs}{the list of CoRecMotif objects to filter.}

\item{seed_name}{a character vector specifying the seed names (probe sets) to
keep.}

\item{pbm_condition}{a character vector specifying the PBM conditions to
keep.}

\item{array_id}{a character vector specifying the array IDs to keep.}

\item{rolling_ic}{a single number specifying the minimum rolling IC to keep.}

\item{motif_strength}{a single number specifying the minimum motif strength
to keep.}

\item{seed_sequence}{a character vector specifying the seed sequences to
keep.}

\item{motif_name}{a character vector specifying the motif names to keep.}

\item{match_name}{a character vector specifying the matching motif names to
keep.}

\item{match_altname}{a character vector specifying the matching motif
altnames to keep.}

\item{match_pvalue}{a single number specifying the maximum match p-value to
keep.}

\item{match_cluster}{a character vector specifying the matching clusters to
keep.}
}
\value{
A list of \linkS4class{CoRecMotif} objects that pass the filters.
}
\description{
Filter a list of \linkS4class{CoRecMotif} objects based on the contents of
\linkS4class{CoRecMotif} slots.
}
\examples{
# Load example CoRecMotifs
corecmotifs <-
    readRDS(
        "example_data/output/example_rep1_v1_a11_run1_all_corecmotifs.rds"
    )

# Filter by rolling IC and motif strength
filtered_corecmotifs <-
    filter_corecmotifs(
        corecmotifs,
        rolling_ic = 1,
        motif_strength = 1
    )
}
