% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_match.R
\name{find_match}
\alias{find_match}
\title{Match a list of \linkS4class{CoRecMotif} objects to reference motifs}
\usage{
find_match(
  corecmotifs,
  reference_motifs_file,
  cluster_assignments = NULL,
  min_overlap = 5,
  meme_path = "/share/pkg.7/meme/5.3.3/install/bin/"
)
}
\arguments{
\item{corecmotifs}{the list of CoRecMotif objects to match to reference
motifs.}

\item{reference_motifs_file}{the path to a MEME format file of reference
motifs to match to.}

\item{cluster_assignments}{a data frame mapping the reference motifs to motif
clusters or NULL to skip the cluster assignment step.}

\item{min_overlap}{a single positive integer specifying the minimum amount of
overlap to require when comparing a CoRecMotif to a reference motif.}

\item{meme_path}{the path to "meme/bin/" or NULL to rely on
\link{\code{memes::runTomTom()}} to find it.}
}
\value{
A list of \linkS4class{CoRecMotif} objects with the match_motif,
match_pvalue, and (optionally) match_cluster slots filled in.
}
\description{
Identifies the reference motif that is the best match to each
\linkS4class{CoRecMotif}.
}
\details{
The PPM representation of each CoRecMotif is compared to the reference motifs
with \link{\code{memes::runTomTom()}} using Euclidean distance ("ed") as the
distance metric. The match_motif and match_pvalue slots are filled in based
on the best match returned by \link{\code{memes::runTomTom()}}. The
match_pvalue slot is corrected for multiple comparisons by multiplying the
raw match p-value by the number of motifs in the reference database. The
match_cluster slot is filled in based on the user-provided cluster
assignments or left empty if no cluster assignments are provided.
}
\examples{
# Load example CoRecMotifs
corecmotifs <-
    readRDS(
        "example_data/output/example_rep1_v1_a11_run1_all_corecmotifs.rds"
    )

# Compare the CoRecMotifs to the reference motifs
matched_corecmotifs <-
    find_match(
        corecmotifs,
        "example_data/Homo_sapiens_JASPAR2022_CORE_filtered.meme"
    )
}
