% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_to_zscore_table.R
\name{fluorescence_to_zscore_table}
\alias{fluorescence_to_zscore_table}
\title{Transform fluorescence values into z-scores}
\usage{
fluorescence_to_zscore_table(
  fluorescence_table,
  fluorescence_columns,
  output_file = NULL
)
}
\arguments{
\item{fluorescence_table}{a data frame of fluorescence values and annotations
for each probe. See 'Details' of \code{\link{annotate_fluorescence_table}}
for a description of the expected annotation columns.}

\item{fluorescence_columns}{a character vector specifying the names of the
columns of \code{fluorescence_table} that contain fluorescence data.}

\item{output_file}{the path to the TSV file where the annotated z-score table
will be written. If NULL (the default), no file is written.}
}
\value{
A data frame of column-wise fluorescence z-scores and probe
annotations.
}
\description{
Converts an annotated data frame of fluorescence data into z-scores.
}
\details{
Raw fluorescence values are normalized against the fluorescence values of the
background probes. Specifically, all the fluorescence values for a given PBM
condition are log-transformed. Then a z-score-like statistic is calculated
using the mean and standard deviation of the fluorescence values of the
background probes in that condition.
}
\examples{
# Load the example fluorescence data
fluorescence_table <-
    read.table(
        "example_data/output/example_rep1_v1_a11_run1_fluorescence.tsv",
        header = TRUE,
        sep = "\t",
        stringsAsFactors = FALSE
    )

# Convert the fluorescence data into z-scores
zscore_table <-
    fluorescence_to_zscore_table(
        fluorescence_table,
        fluorescence_columns = c(
            "UT_SUDHL4_SWISNF_mix",
            "UT_SUDHL4_HDAC1_mix",
            "UT_SUDHL4_PRMT5",
            "UT_SUDHL4_JMJD2A"
        )
    )
}
