% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_fluorescence_to_zscore_table.R
\name{fluorescence_to_zscore_table}
\alias{fluorescence_to_zscore_table}
\title{Transform fluorescence values into z-scores}
\usage{
fluorescence_to_zscore_table(
  fluorescence_table,
  fluorescence_columns,
  output_file = NULL
)
}
\arguments{
\item{fluorescence_table}{a data frame of fluorescence values and annotations
for each probe. See 'Details' of \code{\link[=annotate_fluorescence_table]{annotate_fluorescence_table()}} for a
description of the expected annotation columns.}

\item{fluorescence_columns}{a character vector specifying the names of the
columns of \code{fluorescence_table} that contain fluorescence data.}

\item{output_file}{the path to the TSV file where the annotated z-score table
will be written. If NULL, no file is written. (Default: NULL)}
}
\value{
A data frame of column-wise fluorescence z-scores and probe
annotations.
}
\description{
Converts an annotated data frame of fluorescence data into z-scores.
}
\details{
Raw fluorescence values are normalized against the fluorescence values of the
background probes. Specifically, all the fluorescence values for a given PBM
condition are log-transformed. Then a z-score-like statistic is calculated
using the mean and standard deviation of the fluorescence values of the
background probes in that condition.
}
\examples{
print("FILL THIS IN")
}
