% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_comparisons.R
\name{identify_motif_match}
\alias{identify_motif_match}
\title{Title}
\usage{
identify_motif_match(
  corec_motif,
  reference_motifs_file,
  cluster_assignments = NULL,
  min_overlap = 5,
  method = c("ed", "allr", "pearson", "sandelin", "kullback", "blic1", "blic5", "llr1",
    "llr5"),
  meme_path = "/share/pkg.7/meme/5.3.3/install/bin/"
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Identifies the reference motif that is the best match to the corecmotif and
calculates a p-value for that match.
}
