% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pipeline.R
\name{make_corecmotifs}
\alias{make_corecmotifs}
\title{Make a list of corecmotif objects}
\usage{
make_corecmotifs(
  fluorescence_file,
  pbm_conditions,
  annotation_file,
  output_directory = NULL,
  output_base_name = NULL,
  array_id = NULL
)
}
\arguments{
\item{fluorescence_file}{the path to the file containing the fluorescence
data to load. See 'Details' of \code{\link{annotate_fluorescence_table}}
for expected columns.}

\item{pbm_conditions}{a character vector specifying the PBM conditions (e.g.,
cell type, treatment, and factor profiled) in the order they appear in
\code{fluorescence_file}.}

\item{annotation_file}{the path to the file containing the probe annotations
to use. See 'Details' of \code{\link{annotate_fluorescence_table}} for
expected columns.}

\item{output_directory}{the directory where output files will be saved. No
output files will be created unless output_directory or output_base_name is
provided.}

\item{output_base_name}{the base name for output files. No output files will
be created unless output_directory or output_base_name is provided.}

\item{array_id}{an optional (but recommended) tag specifying the particular
array/experiment the fluorescence data is from.}
}
\value{
A list of \linkS4class{corecmotif} objects, one for each possible
combination of the probe sets annotated in \code{annotation_file} and the
PBM conditions listed in \code{pbm_conditions}.
}
\description{
Creates a list of corecmotif objects using fluorescence data from a CoRec
experiment.
}
\details{
This is a convenience function that calls
\code{\link{annotate_fluorescence_table}},
\code{\link{fluorescence_to_zscore_table}}, and
\code{\link{zscore_table_to_corecmotifs}} and optionally saves the outputs
using file names generated from user input. By default no output files are
created. To save output files, you must supply \code{output_directory},
\code{output_base_name}, or both. If \code{output_directory} is provided but
\code{output_base_name} is not, the files will be saved in the provided
directory using the base name "output". If \code{output_base_name} is
provided but \code{output_directory} is not, the files will be saved in the
current working directory using the provided base name. If \code{array_id} is
provided, it will be appended to \code{output_base_name}. The outputs of
\code{\link{annotate_fluorescence_table}},
\code{\link{fluorescence_to_zscore_table}}, and
\code{\link{zscore_table_to_corecmotifs}} will be saved with the suffixes
"_fluorescence.tsv", "_zscores.tsv", and "_all_corecmotifs.rds" respectively.
}
\examples{
corecmotifs <-
    make_corecmotifs(
        fluorescence_file = "./example_data/hTF_v1_example_fluorescence.dat",
        pbm_conditions = c(
            "UT_SUDHL4_SMARCA4MIX",
            "UT_SUDHL4_HDAC1MIX",
             "UT_SUDHL4_SUZ12",
             "UT_SUDHL4_PRMT5"
        ),
        annotation_file = "example_data/hTF_v1_annotation.tsv",
        output_directory = "./example_data/example_output",
        output_base_name = "example",
        array_id = "v1_a6_run1"
    )
}
\seealso{
\link{annotate_fluorescence_table},
\link{fluorescence_to_zscore_table}, \link{zscore_table_to_corecmotifs}
}
