% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_zscore_table_to_corecmotifs.R
\name{make_zscore_motif}
\alias{make_zscore_motif}
\title{Create a z-score motif}
\usage{
make_zscore_motif(zscore_table, probe_set_name, pbm_condition)
}
\arguments{
\item{zscore_table}{\code{data.frame}. An annotated table of fluorescence
z-scores. See \link{hTF_v1_annotation} for expected annotation columns.}

\item{probe_set_name}{\code{character(1)}. The name of the probe set to use.}

\item{pbm_condition}{\code{character(1)}. The name of the PBM condition to use.}
}
\value{
A matrix with rows corresponding to nucleotides and columns
corresponding to positions in the motif. Each cell is filled with the PBM
condition-specific z-score of the probe with that nucleotide at that
position.
}
\description{
Creates a z-score motif for a given probe set and PBM condition. The rows
correspond to nucleotides and the columns correspond to positions in the
motif.
}
\examples{
print("FILL THIS IN")
}
