% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corecmotif.R
\name{plot_corecmotif}
\alias{plot_corecmotif}
\title{Plot CoRecMotifs and their matching reference motifs}
\usage{
plot_corecmotif(
  corecmotif,
  corecmotif_logo_type = c("delta_zscore", "ICM", "PWM", "PPM", "none"),
  reference_logo_type = c("ICM", "PWM", "PPM", "none"),
  reverse_complement = FALSE,
  correct_orientation = TRUE,
  fade_corecmotif = FALSE,
  fade_reference = FALSE,
  check_corecmotif = TRUE
)
}
\arguments{
\item{corecmotif}{\link[=CoRecMotif-class]{CoRecMotif}. The
\link[=CoRecMotif-class]{CoRecMotif} to plot.}

\item{corecmotif_logo_type}{\code{character(1)}. One of "delta_zscore", "ICM",
"PWM", "PPM", or "none". The type of logo to plot for the
\link[=CoRecMotif-class]{CoRecMotif}. (Default: "delta_zscore")}

\item{reference_logo_type}{\code{character(1)}. One of "ICM", "PWM", "PPM", or
"none". The type of logo to plot for the matching reference motif.
(Default: "ICM")}

\item{reverse_complement}{\code{logical(1)}. Should the CoRecMotif be reversed?
(Default: False)}

\item{correct_orientation}{\code{logical(1)}. Should the reference motif be
reversed if necessary to match the CoRecMotif's orientation? (Default:
TRUE)}

\item{fade_corecmotif}{\code{logical(1)}. Should the
\link[=CoRecMotif-class]{CoRecMotif} logo be faded? (Default: FALSE)}

\item{fade_reference}{\code{logical(1)}. Should the reference motif logo be faded?
(Default: FALSE)}

\item{check_corecmotif}{\code{logical(1)}. Should \code{corecmotif} be checked for
validity? Setting this to FALSE can increase speed, but if \code{corecmotif} is
not a valid \link[=CoRecMotif-class]{CoRecMotif}, it may produce uninformative
error messages. (Default: TRUE)}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a motif logo plot of a \link[=CoRecMotif-class]{CoRecMotif} and/or its
matching reference motif. There are several types of logos to choose from.
}
\examples{
print("FILL THIS IN")
}
