% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_plots.R
\name{plot_corecmotif}
\alias{plot_corecmotif}
\title{Title}
\usage{
plot_corecmotif(
  corec_motif,
  motif_score_type = c("seed_probe_zscore", "mean_highest_zscore",
    "mean_information_content"),
  motif_score_threshold = NA
)
}
\arguments{
\item{motif_score_threshold}{}
}
\value{

}
\description{
Title
}
