% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corecmotif.R
\name{plot_corecmotif}
\alias{plot_corecmotif}
\title{Plot CoRecMotifs and their matching reference motifs}
\usage{
plot_corecmotif(
  corecmotif,
  corecmotif_logo_type = c("delta_zscore", "ICM", "PWM", "PPM", "none"),
  reference_logo_type = c("ICM", "PWM", "PPM", "none")
)
}
\arguments{
\item{corecmotif}{\link[=CoRecMotif-class]{CoRecMotif}. The
\link[=CoRecMotif-class]{CoRecMotif} to plot.}

\item{corecmotif_logo_type}{\code{character(1)}. One of "delta_zscore", "ICM",
"PWM", "PPM", or "none". The type of logo to plot for the
\link[=CoRecMotif-class]{CoRecMotif}. (Default: "delta_zscore")}

\item{reference_logo_type}{\code{character(1)}. One of "ICM", "PWM", "PPM", or
"none". The type of logo to plot for the matching reference motif.
(Default: "ICM")}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a motif logo plot of a \link[=CoRecMotif-class]{CoRecMotif} and/or its
matching reference motif. There are several types of logos to choose from.
}
\examples{
print("FILL THIS IN")
}
