% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_run_full_analysis.R
\name{process_corecmotifs}
\alias{process_corecmotifs}
\title{Filter a list of CoRecMotifs and match them to reference motifs}
\usage{
process_corecmotifs(
  corecmotifs,
  reference_motifs_file,
  cluster_assignments = NULL,
  meme_path = NULL,
  rolling_ic = 1,
  motif_strength = 1,
  n_replicates = 2,
  eucl_distance = 0.4,
  min_overlap = 5,
  match_pvalue = 0.05,
  output_directory = NULL,
  output_base_name = NULL
)
}
\arguments{
\item{corecmotifs}{The list of \link[=CoRecMotif-clas]{CoRecMotifs} to process.}

\item{reference_motifs_file}{the path to a MEME format file of reference
motifs to match to.}

\item{cluster_assignments}{a data frame mapping the reference motifs to motif
clusters or NULL to skip the cluster assignment step.}

\item{meme_path}{the path to "meme/bin/" or NULL to rely on
\code{\link[memes:runTomTom]{memes::runTomTom()}} to find it.}

\item{rolling_ic}{a single number specifying the minimum rolling IC to keep.
(Default: NULL)}

\item{motif_strength}{a single number specifying the minimum motif strength
to keep. (Default: NULL)}

\item{n_replicates}{a single positive integer specifying the minimum number
of replicates to require.}

\item{eucl_distance}{a single number specifying the maximum allowable
Euclidean distance between replicate motifs or NULL to skip the replicate
comparison step.}

\item{min_overlap}{a single positive integer specifying the minimum amount of
overlap to require when comparing a \link[=CoRecMotif-class]{CoRecMotif} to a
reference motif.}

\item{match_pvalue}{a single number specifying the maximum match p-value to
keep. (Default: NULL)}

\item{output_directory}{the directory where output files will be saved. No
output files will be created unless output_directory or output_base_name is
provided. (Default: NULL)}

\item{output_base_name}{the base name for output files. No output files will
be created unless output_directory or output_base_name is provided.
(Default: NULL)}
}
\value{
A filtered list of replicated \link[=CoRecMotif-class]{CoRecMotifs} that
match a reference motif.
}
\description{
Removes \link[=CoRecMotif-class]{CoRecMotifs} with low z-scores or motif strengths
and CoRecMotifs that do not replicate, then compares the remaining
CoRecMotifs to a database of reference motifs to identify the best match.
This is a convenience function that calls \code{\link[=filter_corecmotifs]{filter_corecmotifs()}},
\code{\link[=check_replicates]{check_replicates()}}, \code{\link[=find_match]{find_match()}}, then \code{\link[=filter_corecmotifs]{filter_corecmotifs()}} and
\code{\link[=check_replicates]{check_replicates()}} again, and finally \code{\link[=summarize_corecmotifs]{summarize_corecmotifs()}}.
}
\details{
By default no output files are created. To save output files, you must
provide \code{output_directory}, \code{output_base_name}, or both. The output files are
named according to the following rules.
\itemize{
\item If \code{output_directory} is provided but \code{output_base_name} is not, the files
will be saved in the provided directory using the base name "output".
\item If \code{output_base_name} is provided but \code{output_directory} is not, the files
will be saved in the working directory using the provided base name.
\item The individual output file names will be identified by a suffix appended
to the output base name.
\item The output after the first call to \code{\link[=filter_corecmotifs]{filter_corecmotifs()}} and
\code{\link[=check_replicates]{check_replicates()}} will have the suffix "filtered_corecmotifs.rds".
\item The output of \code{\link[=find_match]{find_match()}} will have the suffix
"matched_corecmotifs.rds".
\item The output after the second call to \code{\link[=filter_corecmotifs]{filter_corecmotifs()}} and
\code{\link[=check_replicates]{check_replicates()}} will have the suffix "significant_corecmotifs.rds".
\item The output of \code{\link[=summarize_corecmotifs]{summarize_corecmotifs()}} will have the suffix
"significant_corecmotifs_summary.tsv".
}
}
\examples{
print("FILL THIS IN")
}
\seealso{
\code{\link[=filter_corecmotifs]{filter_corecmotifs()}}, \code{\link[=check_replicates]{check_replicates()}}, \code{\link[=find_match]{find_match()}},
\code{\link[=summarize_corecmotifs]{summarize_corecmotifs()}}
}
