% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pipeline.R
\name{run_full_analysis}
\alias{run_full_analysis}
\title{Title}
\usage{
run_full_analysis(
  matrix_directory,
  matrix_base_name,
  pbm_conditions_file,
  annotation_file,
  reference_motifs_file,
  run_tag = NA
)
}
\arguments{
\item{run_tag}{}
}
\value{

}
\description{
Title
}
