% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pipeline.R
\name{run_full_analysis}
\alias{run_full_analysis}
\title{Title}
\usage{
run_full_analysis(
  output_directory,
  fluorescence_file,
  pbm_conditions_file = NA,
  pbm_conditions = NA,
  annotation_file,
  reference_motifs_file,
  array_id = NA,
  motif_strength_threshold = 1,
  rolling_ic_threshold = 1.5,
  comparison_method = "ed",
  cluster_assignments_file = NULL,
  pvalue_threshold = 0.05
)
}
\arguments{
\item{array_id}{}
}
\value{

}
\description{
Title
}
