% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_corecmotifs}
\alias{summarize_corecmotifs}
\title{Make a data frame summarizing a list of CoRecMotifs}
\usage{
summarize_corecmotifs(corecmotifs)
}
\arguments{
\item{corecmotifs}{the list of \link[=CoRecMotif-class]{CoRecMotifs} to summarize.}
}
\value{
A data frame with information about a list of
\link[=CoRecMotif-class]{CoRecMotifs}.
}
\description{
Creates a data frame representation of a list of
\link[=CoRecMotif-class]{CoRecMotifs}.
}
\examples{
print("FILL THIS IN")
}
