% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_cluster_match}
\alias{update_cluster_match}
\title{Update the match cluster of a CoRecMotif}
\usage{
update_cluster_match(
  corecmotif,
  cluster_assignments = NULL,
  check_corecmotif = TRUE
)
}
\arguments{
\item{corecmotif}{\link[=CoRecMotif-class]{CoRecMotif}. The
\link[=CoRecMotif-class]{CoRecMotif} to update.}

\item{cluster_assignments}{\code{data.frame} or \code{NULL}. A table mapping the
reference motifs to motif clusters or NULL to reset the \code{match_cluster}
slot to NA. See \link{motif_clusters} for expected columns. (Default: NULL)}

\item{check_corecmotif}{\code{logical(1)}. Should \code{corecmotif} be checked for
validity? Setting this to FALSE can increase speed, but if \code{corecmotif} is
not a valid \link[=CoRecMotif-class]{CoRecMotif}, it may produce uninformative
error messages. (Default: TRUE)}
}
\value{
A \link[=CoRecMotif-class]{CoRecMotif} with its \code{match_cluster} slot
updated.
}
\description{
Updates the \code{match_cluster} slot of a \link[=CoRecMotif-class]{CoRecMotif} based on
the provided cluster assignments and the name of the motif in the
CoRecMotif's \code{match_motif} slot.
}
\examples{
print("FILL THIS IN")
}
