% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_zscore_table_to_corecmotifs.R
\name{zscore_table_to_corecmotifs}
\alias{zscore_table_to_corecmotifs}
\title{Create CoRecMotifs from a table of z-score data}
\usage{
zscore_table_to_corecmotifs(
  zscore_table,
  zscore_columns,
  array_id = NULL,
  output_file = NULL
)
}
\arguments{
\item{zscore_table}{\code{data.frame}. An annotated table of fluorescence
z-scores. See \link{hTF_v1_annotation} for expected annotation columns.}

\item{zscore_columns}{\code{character}. The names of the columns of \code{zscore_table}
that contain fluorescence z-scores.}

\item{array_id}{\code{character(1)} or \code{NULL}. The name of the array/experiment
the z-scores are from. If \code{NULL}, a random ID will be generated.
(Default: NULL)}

\item{output_file}{\code{character(1)} or \code{NULL}. The path to the file where the
results should be saved, or NULL not to save the results. (Default: NULL)}
}
\value{
A list of \link[=CoRecMotif-class]{CoRecMotifs}, one for each possible
combination of the probe sets in \code{zscore_table$probe_set} and the PBM
conditions listed in \code{zscore_columns}.
}
\description{
Converts a table of fluorescence z-scores into a list of
\link[=CoRecMotif-class]{CoRecMotifs}.
}
\details{
This function creates a \link[=CoRecMotif-class]{CoRecMotif} for every possible
combination of the z-score columns listed in \code{zscore_columns} and the
non-background probe sets that appear in the \code{probe_set} column of
\code{zscore_table}. Each column listed in \code{zscore_columns} should contain data
from one PBM condition, and the column name will be used to populate the
\code{pbm_condition} slot of the resulting \link[=CoRecMotif-class]{CoRecMotifs}. The
names of the non-background probe sets found in the \code{probe_set} column will
populate the \code{probe_set} slot, and \code{array_id} (if provided) will populate the
\code{array_id} slot. If no array ID is provided, a random one will be generated
in the format "random_id_xxxxxxxx" where "xxxxxxxx" is replaced with a
randomly generated 8 digit number. The \code{probe_set}, \code{pbm_condition}, and
\code{array_id} slots will be pasted together to create a uniquely identifying
name for each \link[=CoRecMotif-class]{CoRecMotif}.
}
\examples{
print("FILL THIS IN")
}
\seealso{
\link{hTF_v1_annotation} for a description of the probe annotation
columns.
}
