% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_zscore_table_to_corecmotifs.R
\name{zscore_table_to_corecmotifs}
\alias{zscore_table_to_corecmotifs}
\title{Create CoRecMotifs from a table of z-score data}
\usage{
zscore_table_to_corecmotifs(
  zscore_table,
  zscore_columns,
  array_id = NULL,
  output_file = NULL
)
}
\arguments{
\item{zscore_table}{a data frame of z-scores and annotations for each probe.
See 'Details' of \code{\link[=annotate_fluorescence_table]{annotate_fluorescence_table()}} for a description of the
expected annotation columns.}

\item{zscore_columns}{a character vector specifying the names of the columns
of \code{zscore_table} that contain z-score data.}

\item{array_id}{an optional (but recommended) tag specifying the particular
array/experiment the fluorescence data is from. (Default: NULL)}

\item{output_file}{the path to the RDS file where the list of
\link[=CoRecMotif-class]{CoRecMotifs} will be written. If NULL, no file is
written. (Default: NULL)}
}
\value{
A list of \link[=CoRecMotif-class]{CoRecMotifs}, one for each possible
combination of the probe sets in \code{zscore_table} and the PBM conditions
listed in \code{zscore_columns}.
}
\description{
Creates a list of \link[=CoRecMotif-class]{CoRecMotifs} for all possible
combinations of the probe sets present in \code{zscore_table} and the PBM
conditions given in \code{zscore_columns}.
}
\examples{
print("FILL THIS IN")
}
