% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_table_to_corecmotifs.R
\name{zscore_table_to_corecmotifs}
\alias{zscore_table_to_corecmotifs}
\title{Create \linkS4class{corecmotif} objects from a table of z-score data}
\usage{
zscore_table_to_corecmotifs(
  zscore_table,
  zscore_columns,
  output_file = NULL,
  array_id = NULL
)
}
\arguments{
\item{zscore_table}{a data frame of z-scores and annotations for each probe.
See 'Details' of \code{\link{annotate_fluorescence_table}} for a
description of the expected annotation columns.}

\item{zscore_columns}{a character vector specifying the names of the columns
of \code{zscore_table} that contain z-score data.}

\item{output_file}{the path to the RDS file where the list of corecmotif
objects will be written. If NULL (the default), no file is written.}

\item{array_id}{an optional (but recommended) tag specifying the particular
array/experiment the fluorescence data is from.}
}
\value{
A list of \linkS4class{corecmotif} objects, one for each possible
combination of the probe sets in \code{zscore_table} and the PBM conditions
listed in \code{zscore_columns}.
}
\description{
Creates a list of \linkS4class{corecmotif} objects for all possible
combinations of the probe sets present in \code{zscore_table} and the PBM
conditions given in \code{zscore_columns}.
}
\examples{
# Load the example z-score table
zscore_table <-
    read.table(
        "example_data/example_output/example_v1_a11_run1_zscores.tsv",
        header = TRUE,
        sep = "\t",
        stringsAsFactors = FALSE
    )

# Make a list of corecmotif objects
corec_motifs <-
    zscore_table_to_corecmotifs(
        zscore_table,
        zscore_columns = c(
            "v1_a11_run1_UT_SUDHL4_SMARCA4MIX",
            "v1_a11_run1_UT_SUDHL4_HDAC1MIX",
            "v1_a11_run1_UT_SUDHL4_SUZ12",
            "v1_a11_run1_UT_SUDHL4_PRMT5"
        ),
        array_id = "v1_a11_run1"
    )
}
